Fueron modificadas como se muestran los grados

+++ Punto3_version1.0.c
@@ -7,32 +7,32 @@ momentos de un día y determinar:
 #include <stdio.h>
 
 int main(){
-    int temp_1,temp_2,temp_3,prom_; // temp = temperatura & prom_ = promedio
+    float temp_1,temp_2,temp_3,prom_; // temp = temperatura & prom_ = promedio
 
     printf("\nIngresar primer registro de temperatura: ");
-    scanf("%d",&temp_1);
+    scanf("%f",&temp_1);
     
     printf("\nIngresar segundo registro de temperatura: ");
-    scanf("%d",&temp_2);
+    scanf("%f",&temp_2);
 
     printf("\nIngresar tercer registro de temperatura: ");
-    scanf("%d",&temp_3);
+    scanf("%f",&temp_3);
 
     //a. Promedio de temperatura
     prom_ = (temp_1 + temp_2 + temp_3) / 3;
 
-    printf("\nEl promedio de temperatura es: %d",prom_);
+    printf("\nEl promedio de temperatura es: %.2f",prom_);
     
     //b. Temperatura mayor al promedio
     if (temp_1>prom_)
     {
-        printf("\nTemperatura de %d es mayor al promedio de %d",temp_1,prom_);
+        printf("\nTemperatura de %.1f es mayor al promedio de %.1f",temp_1,prom_);
     }else if (temp_2>prom_) // Se llama condición anidada
         {
-            printf("\nTemperatura de %d es mayor al promedio de %d",temp_2,prom_);
+            printf("\nTemperatura de %.1f es mayor al promedio de %.1f",temp_2,prom_);
         }else
             {
-                printf("\nTemperatura de %d es mayor al promedio de %d",temp_3,prom_);
+                printf("\nTemperatura de %.1f es mayor al promedio de %.1f",temp_3,prom_);
             }
     
     return 0;
